requirements.txt
testinfra==1.6.5
PyHamcrest==1.9.0